+++ src/JS/markup.js
@@ -1,4 +1,4 @@
-//
+// make request markup
 
 function sectionRequestMarkup() {}
 function searchRequestMarkup() {}
